Windows build projects
This is currently not working for some reason. I'm getting linking
errors from within the libwebp library I'm building. I get these errors
even if I import the files into the main, which is really weird.

My workaround (not in this version of the project) is to download the
library built by Google, which links fine. It's currently only available
in 32-bit though.