Switched over to libwebm
A lot of work went into this changeset and a lot of work is still to
go. Dropped nesegg and whever else in favor of the real deal, libwebm.
Put the basics for Vorbis audio exporting in, but then commented it all
out because it's not working properly. Have to put this project aside
for the moment.