First working version
Huzzah! It finally works!

Yes yes yes yes yes yes yes yes yes yes yes!!!!!!!!!!!!!!!

Sorry, but this was a tough one. Still plenty of work to do, but it
does now seem to read and write audio and video that appears to play
back according to plan.

There are surely bugs left to be found and an interface left to to
design, but for now I can smile. I feel like this is the hardest
programming thing I've had to do. That I had to go the longest time to
get to something that actually worked. The first plug-in I ever wrote
(SuperPNG) may not have taken so long to get up and running.

While working on this thing I took a "break" and pushed out a
mostly-complete WebP plug-in, which was easy by comparison. So just
give me one more second to exhale.

Hooray!!!!!!!!!!!!