Tried continuous audio reading, but turned it off
When I start reading audio in a packet, I use the timestamp it gives
me, even though this timestamp often seems wrong and rounded off. For
frames it's close enough, but it's tough to know the exact audio sample
to start with in this situation. Between reads, the mismatch may cause
the reads not to properly align.

So I thought I'd try remembering the last sample I read and then pick
up right where I left off when Premiere asked for more samples
immediately following the previous read. Genius, right?

Except the audio seems to drift when I do this. By always re-aligning
with the packet time I may get gaps but what I do have is lines up
right. I've turned it off for now, but leaving the code in place for
later exploration.