Fix problem with constant quality
I can't tell what effect VP8E_SET_CQ_LEVEL has, but apparently it needs
to be between rc_min_quantizer and rc_max_quantizer. It defaults to 10,
so if max_q got below that I'd get an error. This would happen when
quality was set above 93 or so. This fix is for issue #5.