Init audio encoders only once
Move the initialization functions so they only happen once when the
file is opened.

This seems to be working, which must mean that the decoders are never
called on separate threads simultaneously. This seems to be working,
although I did get a crash at one point.I haven't been able to
reproduce, but I'll be careful and leave this on a branch for a while
until I'm really confident it's working.