Update Windows build files
Bugs in issue #8 were caused by our manual way of updating the Windows
build files.

To do it properly we have to build everything on the command line with
./configure --target=x86_64-win64-vs9 --enable-static-msvcrt in Cygwin
and then `make` from the Visual Studio shell.

Then we have to convert the paths in the vcproj files from wherever we
had libvpx to the relative paths we really want.